+++ mlir/lib/Conversion/ArithAndMathToAPFloat/MathToAPFloat.cpp
@@ -6,9 +6,9 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "mlir/Conversion/MathToAPFloat/MathToAPFloat.h"
 #include "Utils.h"
 
+#include "mlir/Conversion/ArithAndMathToAPFloat/MathToAPFloat.h"
 #include "mlir/Dialect/Func/IR/FuncOps.h"
 #include "mlir/Dialect/Func/Utils/Utils.h"
 #include "mlir/Dialect/Math/IR/Math.h"
